% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitart.R
\name{pitart}
\alias{pitart}
\title{F1 pitstop ASCII art}
\format{
ASCII string
}
\usage{
pitart(n = 1)
}
\arguments{
\item{n}{Integer. ASCII pit stop to generate. From 1 (by default) to 5}
}
\value{
A string containing the ASCII art of a F1 pit stop
}
\description{
Funny ASCII F1 pitstop for title_text argument in pitplot() function
}
\examples{
pitart(1)
pitart(2)
pitart(3)
pitart(5)
}
