% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchamp.R
\name{pitchamp}
\alias{pitchamp}
\title{Pit stop standings by season}
\usage{
pitchamp(year)
}
\arguments{
\item{year}{Season pit stop standings (integer). 2015 or higher}
}
\value{
A tibble containing the pit stop season standings points (or wins in 2015 and 2016)
}
\description{
Pit stop official standings by season (since 2015)
}
\examples{
\donttest{
pitchamp(2015)
}
}
